Fix: small icon exception
