+++ .idea/libraries/Gradle__androidx_core_core_1_5_0_aar.xml
@@ -2,15 +2,11 @@
   <library name="Gradle: androidx.core:core:1.5.0@aar">
     <ANNOTATIONS>
       <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/e98c05ceb0fb106bfccdec2dea6fe38d/transformed/core-1.5.0/annotations.zip!/" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/e98c05ceb0fb106bfccdec2dea6fe38d/transformed/core-1.5.0/annotations.zip!/" />
     </ANNOTATIONS>
     <CLASSES>
       <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/e98c05ceb0fb106bfccdec2dea6fe38d/transformed/core-1.5.0/jars/classes.jar!/" />
       <root url="file://$USER_HOME$/.gradle/caches/transforms-3/e98c05ceb0fb106bfccdec2dea6fe38d/transformed/core-1.5.0/res" />
       <root url="file://$USER_HOME$/.gradle/caches/transforms-3/e98c05ceb0fb106bfccdec2dea6fe38d/transformed/core-1.5.0/AndroidManifest.xml" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/e98c05ceb0fb106bfccdec2dea6fe38d/transformed/core-1.5.0/AndroidManifest.xml" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/e98c05ceb0fb106bfccdec2dea6fe38d/transformed/core-1.5.0/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/e98c05ceb0fb106bfccdec2dea6fe38d/transformed/core-1.5.0/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
