+++ .idea/misc.xml
@@ -18,12 +18,16 @@
         <entry key="..\:/Android_projects/NiceFilmApp/app/src/main/res/mipmap-anydpi-v26/ic_home.xml" value="0.22135416666666666" />
         <entry key="..\:/Android_projects/NiceFilmApp/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.3828125" />
         <entry key="..\:/Android_projects/NiceFilmApp/app/src/main/res/mipmap-anydpi-v26/ic_launcher_round.xml" value="0.22135416666666666" />
+        <entry key="..\:/android/NiceFilmApp/app/src/main/res/drawable/ic_home_svgrepo_com.xml" value="0.2074074074074074" />
         <entry key="..\:/android/NiceFilmApp/app/src/main/res/layout/activity_main.xml" value="0.20520833333333333" />
         <entry key="..\:/android/NiceFilmApp/app/src/main/res/layout/film_recycler_view_item.xml" value="0.20520833333333333" />
         <entry key="..\:/android/NiceFilmApp/app/src/main/res/layout/fragment_current_film.xml" value="0.36666666666666664" />
         <entry key="..\:/android/NiceFilmApp/app/src/main/res/layout/fragment_current_film_maps.xml" value="0.27807971014492755" />
         <entry key="..\:/android/NiceFilmApp/app/src/main/res/layout/fragment_main_films.xml" value="0.26132246376811596" />
         <entry key="..\:/android/NiceFilmApp/app/src/main/res/layout/fragment_settings.xml" value="0.33" />
+        <entry key="..\:/android/NiceFilmApp/app/src/main/res/layout/qwe.xml" value="0.2015625" />
+        <entry key="..\:/android/NiceFilmApp/app/src/main/res/layout/ref.xml" value="0.2015625" />
+        <entry key="..\:/android/NiceFilmApp/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.22135416666666666" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.29270833333333335" />
         <entry key="app/src/main/res/layout/fragment_film_card.xml" value="0.29270833333333335" />
         <entry key="app/src/main/res/layout/fragment_main_films.xml" value="0.29270833333333335" />
