+++ app/build.gradle
@@ -36,9 +36,11 @@ android {
     productFlavors {
         free {
             dimension 'version'
+            versionNameSuffix '.free'
         }
         paid {
             dimension 'version'
+            versionNameSuffix '.paid'
         }
     }
 }
